- Moved the isValidStringCharacter to the JottTokenizer
